chore: move assertNonNullish to utils
