+++ src/lib/htmlElements.ts
@@ -1,12 +1,4 @@
-export function assertNonNullish<T>(
-  value: T,
-  message: string,
-): asserts value is NonNullable<T> {
-  if (value === null || value === undefined) {
-    throw Error(message);
-  }
-  return;
-}
+import { assertNonNullish } from "./utils";
 
 export function htmlElementNotFoundMessage(id: string) {
   return `HTMLElement #${id} not found!`;
