+++ src/lib/utils.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { roundToTwoDecimals } from "./utils";
+import { assertNonNullish, roundToTwoDecimals } from "./utils";
 
 describe("roundToTwoDecimals", () => {
   describe("given a number", () => {
@@ -11,5 +11,22 @@ describe("roundToTwoDecimals", () => {
       expect(roundToTwoDecimals(0)).toBe(0);
       expect(typeof roundToTwoDecimals(8.123)).toBe("number");
     });
+  });
+});
+
+describe("F", () => {
+  it("throws an error when the value is null or undefined", () => {
+    expect(() => assertNonNullish(null, "Value cannot be null")).toThrowError(
+      "Value cannot be null",
+    );
+    expect(() =>
+      assertNonNullish(undefined, "Value cannot be undefined"),
+    ).toThrowError("Value cannot be undefined");
+  });
+
+  it("does not throw an error when the value is not null or undefined", () => {
+    assertNonNullish(1, "Value cannot be null");
+    assertNonNullish("test", "Value cannot be undefined");
+    assertNonNullish({}, "Value cannot be null or undefined");
   });
 });
